# Specialize `tyView`