Use specialized matching before calling `splitTyAppM` to get a more efficient case-tree.

Before:
```
benchmarking normalization of tests/shouldwork/Basic/AES.hs
time                 323.7 ms   (306.5 ms .. 355.1 ms)
                     0.996 R²   (0.985 R² .. 1.000 R²)
mean                 319.7 ms   (315.7 ms .. 327.6 ms)
std dev              8.781 ms   (93.23 μs .. 10.84 ms)
variance introduced by outliers: 16% (moderately inflated)
```

After:
```
benchmarking normalization of tests/shouldwork/Basic/AES.hs
time                 295.3 ms   (266.7 ms .. 330.1 ms)
                     0.995 R²   (0.975 R² .. 1.000 R²)
mean                 293.3 ms   (287.9 ms .. 303.8 ms)
std dev              8.808 ms   (513.4 μs .. 11.32 ms)
variance introduced by outliers: 16% (moderately inflated)
```