@christiaanb 2020-02-20
A 10% speed up is pretty good imo. Could you elaborate on how you’d structure the worker-wrapper?